Fixed possible out-of-bounds bug
Adding the "w" command requires the buffer to be consistent at all times.
This was not the case when user attempts to write more than entire buffer.
The old code returned an error leaving the buffer full. A subsequent write-only
call could lead to out-of-bounds. Fixed by adding 0 to last element on buffer write
failure.